+++ bindings/SofaTypes/src/SofaPython3/SofaTypes/Binding_Vec.h
@@ -34,7 +34,7 @@ using namespace pybind11::literals;
 using sofa::defaulttype::Vec;
 
 namespace pyVec {
-template <int N, class T>
+template <size_t N, class T>
 std::string __str__(const Vec<N, T> &self, bool repr = false);
 } // namespace pyVec
